+++ src/main/java/com/maxmind/minfraud/request/Transaction.java
@@ -21,7 +21,7 @@ public final class Transaction extends AbstractModel {
     private final Shipping shipping;
     private final List<ShoppingCartItem> shoppingCart;
 
-    protected Transaction(Transaction.Builder builder) {
+    private Transaction(Transaction.Builder builder) {
         account = builder.account;
         billing = builder.billing;
         creditCard = builder.creditCard;
@@ -48,7 +48,7 @@ public final class Transaction extends AbstractModel {
         Order order;
         Payment payment;
         Shipping shipping;
-        List<ShoppingCartItem> shoppingCart = new ArrayList<>();
+        final List<ShoppingCartItem> shoppingCart = new ArrayList<>();
 
         /**
          * The constructor for {@code Builder}